requirements.txt
APScheduler
attrs
beautifulsoup4
blinker
certifi
cffi
charset-normalizer
click
colorama
Flask
h11
idna
itsdangerous
Jinja2
lxml
MarkupSafe
outcome
pycparser
pycryptodome
PySocks
pytz
requests
selenium
setuptools
six
sniffio
sortedcontainers
soupsieve
trio
trio-websocket
typing_extensions
tzdata
tzlocal
undetected-chromedriver
urllib3
websocket-client
websockets
Werkzeug
wsproto
